+++ scout/commands/load/report.py
@@ -4,7 +4,7 @@ import logging
 import click
 from flask.cli import with_appcontext
 
-from scout.load.report import load_delivery_report, load_cnv_report
+from scout.load.report import load_delivery_report, load_cnv_report, load_coverage_qc_report
 from scout.server.extensions import store
 
 LOG = logging.getLogger(__name__)
@@ -45,6 +45,29 @@ def cnv_report(case_id, report_path, update):
 
     try:
         load_cnv_report(
+            adapter=adapter,
+            case_id=case_id,
+            report_path=report_path,
+            update=update,
+        )
+        LOG.info("saved report to case!")
+    except Exception as err:
+        LOG.error(err)
+        raise click.Abort()
+
+
+@click.command("coverage-qc-report")
+@click.argument("case-id", required=True)
+@click.argument("report-path", type=click.Path(exists=True), required=True)
+@click.option("-u", "--update", is_flag=True, help="update coverage and qc report for a sample")
+@with_appcontext
+def coverage_qc_report(case_id, report_path, update):
+    """Add coverage and qc report to an existing case."""
+
+    adapter = store
+
+    try:
+        load_coverage_qc_report(
             adapter=adapter,
             case_id=case_id,
             report_path=report_path,